Use a stand-in value for the platform when disconnected
Some operations do not involve a platform, and should be able to
proceed when the daemon is not connected. To decouple those things,
instead of returning an error when the platform is disconnected,
return an object that returns errors on being used.

Tidied up some of the logging (and fmt.Printf) while I'm here.